test debugger implemented (tmp)

+++ include/var/_string.h
@@ -61,7 +61,7 @@ public:
 
 	String upper() const;
 	String lower() const;
-	String substr(size_t p_start, size_t p_end) const;
+	String substr(int64_t p_start, int64_t p_end) const;
 	bool endswith(const String& p_str) const;
 	bool startswith(const String& p_str) const;
 	Array split(const String& p_delimiter = "") const;
